Fix betweenness label indentation bug and land ShareGNN invariant-layer optimizations
save_betweenness_centrality_labels had its body de-indented out of the
"if not file.exists()" guard, so start_time was only bound when the label
file was missing while the timing code ran unconditionally. The bare except
turned the resulting UnboundLocalError into a misleading "No save time path
given" ValueError, crashing preprocessing in examples/share_gnn_basic.

Also includes the pending ShareGNN invariant-layer optimization work
(specs/07, specs/08), its integration tests, and shared test fixtures.

+++ src/simplegnn/framework/core.py
@@ -1125,7 +1125,7 @@ def preprocess_graph_data(experiment_configuration: dict):
                                 input_features=experiment_configuration.get('input_features', None),
                                 output_features=experiment_configuration.get('output_features', None),
                                 graph_format=experiment_configuration.get('format', 'RuleGNNDataset'),
-                                precision=experiment_configuration.get('precision', 'double'),
+                                precision=experiment_configuration.get('precision', 'float'),
                                 experiment_config=experiment_configuration)
     # move the dataset to the device
     graph_data.to(experiment_configuration.get('device', 'cpu'))
